Fix #1219: move getProcessesOnPort to a leaf module for the doctor check
`codev doctor` needs the running Tower's PID, and importing it from
`agent-farm/commands/tower.ts` dragged that command's import graph into
doctor.ts — including `utils/shell.ts`, which calls `promisify(exec)` at module
scope. `doctor.test.ts` mocks `node:child_process` without an `exec` export, so
all 24 of its cases failed at import time.

Move the function to `agent-farm/utils/port.ts` (execSync only, no module-scope
side effects) and re-export it from `commands/tower.ts` so existing consumers,
including tower-stop.test.ts, are untouched. doctor.ts takes DEFAULT_TOWER_PORT
straight from `@cluesmith/codev-sdk/constants` for the same reason.
